+++ packages/media-query-list-parser/test/test.mjs
@@ -88,6 +88,7 @@ import './cases/various/0005.mjs';
 import './cases/various/0006.mjs';
 import './cases/various/0007.mjs';
 import './cases/various/0008.mjs';
+import './cases/various/0009.mjs';
 
 import './cases/walker/0001.mjs';