+++ packages/media-query-list-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes to Media Query List Parser
 
+### 2.1.4
+
+_August 5, 2023_
+
+- Do not allow keywords as media feature values in range context queries.
+
 ### 2.1.3
 
 _July 24, 2023_